domenicop's fork first commit
Please note that I know I included too many changes in one commit. It will happen only the first time, I promise.
Added the possibility to select horizontal and vertical padding of the layout inside of the default Crouton separately.
Added responsive padding and text sizes that adapt to phones, 7" tablets and 10" tablets.
Renamed ALERT style to ERROR style.
Added a WARNING style, with a yellow Crouton.
Moved the default styles from Style.java to the new Styles.java class. The styles need to be initialized first now because we need to read responsive padding and text sizes from the library resources to inflate a Crouton layout.
Updated gradle to 0.9.+ and built tools to 19.0.3.
The sample project is not built by default anymore, it was commented out in the settings.gradle file.